ci: switch to cargo install --force rustfmt-nightly
The rustfmt rustup component can fail to build on nightly toolchain
releases.  To avoid this, we install the latest released rustfmt-nightly
crate.  To force recompilation of rustfmt-nightly against nightly (which
is required), we use --cache.

This costs a fair amount of time (about as much as the rustodon build
proper), but until stabilization occurs I'm afraid it's the best we can
do.