database: move to its own crate
This is done so that we can use proc_macros in the main crate,
for Maud templating. Due to a rustc bug (rust-lang/rust#44925),
#![feature(proc_macros)] and custom derive attributes don't play
nice together, so switching on proc_macros breaks Diesel derives.

Thankfully, just splitting the Diesel code out into its own crate
and depending on it works fine, since crates are a feature boundary.