db: Maintain per-thread node IDs (#17)
The Snowflake algorithm requires that each node generating IDs has a
unique ID; flaken requires that this ID be a u64.  Rust threads *do*
have IDs (i.e. the ThreadId struct) but this ID is opaque, and the only
way to get it as a u64 is to hash the ID, which eliminates the
uniqueness guarantee.

To maintain unique IDs for nodes, then, this commit introduces
thread-local integral IDs taken from a counter that's protected by a
mutex.