[WIP] Rewrite of entity parsing to nom
Entity parsing is now done in one pass, instead of multiple passes. This
should be more efficient and guarantees that entities will not overlap.

WIP notes: The current code was naïvely ported from regexes, meaning
that characters are consumed and not avaiable for stuff like last
character checks. These can be worked around (hashtag parsing is the
only working parser right now and can be seen in action from the cli
example), but I wanted to get something pushed.

* Make URL parsing work (backtracking fixes)
* Make mention parsing work
* Remove dead regex module
* Make tests compile
* Make tests pass
* Rewrite usage points in rustodon